深度推定と人物INDEX並び替え処理を分割
