+++ VideoToDepth_en.bat
@@ -30,6 +30,13 @@ IF /I "%OPENPOSE_JSON%" EQU "" (
     EXIT /B
 )
 
+rem ---  深度推定結果ディレクトリパス
+echo If you have already done depth estimation, please enter the full path of the depth estimation result directory. ({Video name}_json_{datetime}_depth)
+echo This setting is available only for half size alphanumeric characters.
+echo If there is a depth file in the directory, the depth estimation result is read from that file.
+set PAST_DEPTH_PATH=
+set /P PAST_DEPTH_PATH=** Depth estimation result directory path: 
+
 rem ---  Depth estimation interval
 echo --------------
 set DEPTH_INTERVAL=10
@@ -110,6 +117,6 @@ IF /I "%IS_DEBUG%" EQU "warn" (
 )
 
 rem ---  python 実行
-python tensorflow/predict_video.py --model_path tensorflow/data/NYU_FCRN.ckpt --video_path %INPUT_VIDEO% --json_path %OPENPOSE_JSON% --interval %DEPTH_INTERVAL% --reverse_specific "%REVERSE_SPECIFIC_LIST%" --order_specific "%ORDER_SPECIFIC_LIST%" --avi_output %AVI_OUTPUT% --verbose %VERBOSE% --number_people_max %NUMBER_PEOPLE_MAX% --end_frame_no %FRAME_END%
+python tensorflow/predict_video.py --model_path tensorflow/data/NYU_FCRN.ckpt --centerz_model_path tensorflow/data2/centerz-depth.ckpt --past_depth_path "%PAST_DEPTH_PATH%" --video_path %INPUT_VIDEO% --json_path %OPENPOSE_JSON% --interval %DEPTH_INTERVAL% --reverse_specific "%REVERSE_SPECIFIC_LIST%" --order_specific "%ORDER_SPECIFIC_LIST%" --avi_output %AVI_OUTPUT% --verbose %VERBOSE% --number_people_max %NUMBER_PEOPLE_MAX% --end_frame_no %FRAME_END%
